Change the supercycle index selector
Formerly it would filter based on index position and
keep state related to that.  Now it filters based on data
and the only state is the currently excluded data.

The new scheme is more robust to external changes in the
base data, either reordering, adding rows, or deleting
them.  It will also transparantly handle changes in configuration
of the supercycle itself.